wstępny podział na komponenty
